requirements-hf.txt
# Ultra-lightweight dependencies using HuggingFace API for embeddings
# Core framework
fastapi==0.128.5
uvicorn[standard]
python-dotenv

# LangGraph & AI
langgraph
langchain-core
langchain-groq
groq

# Vector DB (client only, no local model)
qdrant-client

# HTTP client for HuggingFace API
requests

# Data processing
pydantic>=2.0

# NO sentence-transformers
# NO torch/pytorch
